+++ stag_lib/stag.h
@@ -7,10 +7,9 @@
 
 namespace stag {
   // Provide metadata about the library version
-  static const unsigned VERSION_MAJOR = 1;
-  static const unsigned VERSION_MINOR = 3;
+  static const unsigned VERSION_MAJOR = 2;
+  static const unsigned VERSION_MINOR = 0;
   static const unsigned VERSION_PATCH = 0;
-
 }
 
 #endif //STAG_TEST_STAG_H