+++ test/stag_test.cpp
@@ -9,7 +9,7 @@
 #include <stag.h>
 
 TEST(StagTest, Version) {
-  EXPECT_EQ(stag::VERSION_MAJOR, 1);
-  EXPECT_EQ(stag::VERSION_MINOR, 3);
+  EXPECT_EQ(stag::VERSION_MAJOR, 2);
+  EXPECT_EQ(stag::VERSION_MINOR, 0);
   EXPECT_EQ(stag::VERSION_PATCH, 0);
 }