+++ stag_lib/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [Issue #99](https://github.com/staglibrary/stag/issues/92): Add methods for adding and removing edges
 
 ### Fixed
 - [Issue #126](https://github.com/staglibrary/stag/issues/126): Check Laplacian matrix is diagonally dominant